+++ src/NzbDrone.Api/Wanted/MovieCutoffModule.cs
@@ -26,6 +26,11 @@ namespace NzbDrone.Api.Wanted
 
             pagingSpec.FilterExpression = _movieService.ConstructFilterExpression(pagingResource.FilterKey, pagingResource.FilterValue);
 
+            if (pagingResource.FilterKey == null)
+            {
+                pagingSpec.FilterExpression = m => m.Monitored == true;
+            }
+
             var resource = ApplyToPage(_movieCutoffService.MoviesWhereCutoffUnmet, pagingSpec, v => MapToResource(v, true));
 
             return resource;
