+++ src/NzbDrone.Api/Wanted/MovieMissingModule.cs
@@ -15,9 +15,9 @@ namespace NzbDrone.Api.Wanted
     {
         protected readonly IMovieService _movieService;
 
-        public MovieMissingModule(IMovieService movieService, 
-                                  IQualityUpgradableSpecification qualityUpgradableSpecification, 
-                                  IBroadcastSignalRMessage signalRBroadcaster) 
+        public MovieMissingModule(IMovieService movieService,
+                                  IQualityUpgradableSpecification qualityUpgradableSpecification,
+                                  IBroadcastSignalRMessage signalRBroadcaster)
             : base(movieService, qualityUpgradableSpecification, signalRBroadcaster, "wanted/missing")
         {
 
@@ -30,6 +30,10 @@ namespace NzbDrone.Api.Wanted
             var pagingSpec = pagingResource.MapToPagingSpec<MovieResource, Core.Movies.Movie>("title", SortDirection.Descending);
 
             pagingSpec.FilterExpression = _movieService.ConstructFilterExpression(pagingResource.FilterKey, pagingResource.FilterValue);
+            if (pagingResource.FilterKey != "monitored")
+            {
+                pagingSpec.FilterExpression = m => m.Monitored == true;
+            }
 
             var resource = ApplyToPage(_movieService.MoviesWithoutFiles, pagingSpec, v => MapToResource(v, true));
